refactor : 주석제거 및 잘못 배정된 패키지 변경
