+++ src/test/java/kitchenpos/menu/ui/MenuControllerTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.ui;
 
 import kitchenpos.common.ControllerTest;
+import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.*;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menu.ui.MenuRestController;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
